fixed layout and file structures to make sure the project is clean and logical
