+++ mobile/app/(auth)/_layout.tsx
@@ -1,10 +1,1 @@
-import { Stack } from "expo-router";
-
-export default function AuthLayout() {
-  return (
-    <Stack>
-      <Stack.Screen options={{ headerShown: false }} name="signin" />
-      <Stack.Screen name="signup" />
-    </Stack>
-  );
-}
+export { default } from "@navigation/NVAuthStackLayout";
