+++ mobile/components/controls/PostCardPlaceholder.tsx
