+++ mobile/components/index.ts
@@ -1,6 +1,2 @@
-export { default as PVActionButton } from './controls/PVActionButton';
-export { default as PVIcon } from './controls/PVIcon';
-export { default as PVFormInput } from './controls/PVFormInput';
-export { default as PVFormToggle } from './controls/PVFormToggle';
-export { default as PVImageButton } from './controls/PVImageButton';
-export * from './PostCard';
+export { default as PostCardPlaceholder } from "./controls/PostCardPlaceholder";
+export { default as HomeScreenFeed } from "./controls/HomeScreenFeed";
