+++ mobile/context/RootProvider.tsx
@@ -1,10 +1,12 @@
 import * as Sentry from "@sentry/react-native";
-import { DataProvider } from "@/hooks";
-import { AuthProvider } from "@/hooks";
-import { TranslationProvider } from "@/hooks";
-import { AppLoader } from "./AppLoader";
+
+import { TranslationProvider } from "@context/TranslationProvider";
+
 import { Text } from "react-native";
 import { EnvironmentProvider } from "./EnvironmentProvider";
+import { GestureHandlerRootView } from "react-native-gesture-handler";
+import { AuthProvider, DataProvider } from "@context";
+import { AppProvider } from "./AppProvider";
 
 const SentryFallback = () => (
   <Text style={{ margin: 20 }}>
@@ -12,14 +14,16 @@ const SentryFallback = () => (
   </Text>
 );
 
-export function RootProviders() {
+export function RootProvider() {
   return (
     <EnvironmentProvider>
       <Sentry.ErrorBoundary fallback={SentryFallback} showDialog>
         <DataProvider>
           <TranslationProvider>
             <AuthProvider>
-              <AppLoader />
+              <GestureHandlerRootView style={{ flex: 1 }}>
+                <AppProvider />
+              </GestureHandlerRootView>
             </AuthProvider>
           </TranslationProvider>
         </DataProvider>
