+++ mobile/context/ThemeProvider.tsx
@@ -1,7 +1,7 @@
-import React from 'react';
+import React from "react";
 
-import {light} from '../constants/';
-import {ITheme, IThemeProvider} from '../constants/types';
+import { light } from "../constants";
+import { ITheme, IThemeProvider } from "../constants/types";
 
 export const ThemeContext = React.createContext({
   theme: light,
@@ -14,13 +14,13 @@ export const ThemeProvider = ({
   setTheme = () => {},
 }: IThemeProvider) => {
   return (
-    <ThemeContext.Provider value={{theme, setTheme}}>
+    <ThemeContext.Provider value={{ theme, setTheme }}>
       {children}
     </ThemeContext.Provider>
   );
 };
 
-export default function useTheme(): ITheme {
-  const {theme} = React.useContext(ThemeContext);
+export function useTheme(): ITheme {
+  const { theme } = React.useContext(ThemeContext);
   return theme;
 }
