+++ mobile/hooks/index.ts
@@ -1,11 +1,4 @@
-// hooks index.ts
-export { ThemeContext, ThemeProvider, default as useTheme } from './useTheme';
-export { DataContext, DataProvider, useData } from './useData';
-export { default as useScreenOptions } from './useScreenOptions';
-export {
-  TranslationContext,
-  TranslationProvider,
-  useTranslation,
-} from './useTranslation';
-export * from './useAuth';
-export * from './useScrollDirection';
+export * from "./useColorScheme";
+export * from "./useScreenOptions";
+export * from "./useScrollDirection";
+export * from "./useThemeColor";
