+++ mobile/navigation/NVAuthStackLayout.tsx
@@ -1,0 +1,11 @@
+import { Stack } from "expo-router";
+
+export default function NVAuthStackLayout() {
+  return (
+    <Stack
+      screenOptions={{
+        headerShown: false, // if you want no headers
+      }}
+    />
+  );
+}
