+++ mobile/navigation/NVFriendStackLayout.tsx
@@ -1,0 +1,11 @@
+import { Stack } from "expo-router";
+
+export default function NVFriendStackLayout() {
+  return (
+    <Stack
+      screenOptions={{
+        headerShown: false,
+      }}
+    />
+  );
+}
