+++ mobile/navigation/NVRootLayout.tsx
@@ -1,0 +1,11 @@
+// navigation/NVRootLayout.tsx
+import { RootProvider } from "@context";
+import { Slot } from "expo-router";
+
+export default function NVRootLayout() {
+  return (
+    <RootProvider>
+      <Slot />
+    </RootProvider>
+  );
+}
