+++ mobile/package.json
@@ -9,7 +9,9 @@
     "ios": "expo run:ios",
     "web": "expo start --web",
     "test": "jest --watchAll",
-    "lint": "expo lint"
+    "lint": "expo lint",
+    "generate-barrels": "ts-node scripts/generateBarrels.ts",
+    "watch-barrels": "chokidar 'src/**/*.{ts,tsx}' -c 'npm run generate-barrels'"
   },
   "jest": {
     "preset": "jest-expo"
@@ -67,10 +69,12 @@
     "@types/react": "~18.3.12",
     "@types/react-native-dotenv": "^0.2.2",
     "@types/react-test-renderer": "^18.3.0",
+    "chokidar-cli": "^3.0.0",
     "jest": "^29.2.1",
     "jest-expo": "~52.0.6",
     "react-test-renderer": "18.3.1",
-    "typescript": "^5.3.3"
+    "ts-node": "^10.9.2",
+    "typescript": "^5.8.3"
   },
   "private": true,
   "expo": {
