--- mobile/providers/index.ts
@@ -1,4 +1,0 @@
-// providers/index.ts
-export * from "./AppLoader";
-export * from "./EnvironmentProvider";
-export * from "./RootProviders";
